+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v0.6.0
+
+  * Enhancements
+    * Update Finch to 0.7
+
+  * Breaking changes
+    * Require Elixir 1.7+
+
+## v0.5.0
+
+  * Bug fixes
+    * Fix Mollie.child_spec [#9](https://github.com/jarroput/mollie/pull/9)
+
 ## v0.4.0
 
   * Enhancements